+++ src/test/scala/dev/kamu/engine/spark/ingest/MergeStrategySnapshotTest.scala
@@ -10,7 +10,6 @@ package dev.kamu.engine.spark.ingest
 
 import java.sql.Timestamp
 
-import dev.kamu.core.utils.ManualClock
 import dev.kamu.core.utils.test.KamuDataFrameSuite
 import dev.kamu.engine.spark.ingest.merge.SnapshotMergeStrategy
 import org.scalatest.FunSuite
@@ -29,7 +28,6 @@ case class EmployeeV2(
 )
 
 case class EmployeeEvent(
-  system_time: Timestamp,
   event_time: Timestamp,
   observed: String,
   id: Int,
@@ -38,7 +36,6 @@ case class EmployeeEvent(
 )
 
 case class EmployeeEventV2(
-  system_time: Timestamp,
   event_time: Timestamp,
   observed: String,
   id: Int,
@@ -52,12 +49,6 @@ class MergeStrategySnapshotTest extends FunSuite with KamuDataFrameSuite {
 
   protected override val enableHiveSupport = false
 
-  def clockAt(timestamp: Timestamp) = {
-    val systemClock = new ManualClock()
-    systemClock.set(timestamp)
-    systemClock
-  }
-
   test("From empty") {
     val curr = sc
       .parallelize(
@@ -70,26 +61,25 @@ class MergeStrategySnapshotTest extends FunSuite with KamuDataFrameSuite {
       .toDF()
 
     val t_e = new Timestamp(0)
-    val t_s = new Timestamp(1)
 
     val actual = new SnapshotMergeStrategy(
       primaryKey = Vector("id"),
-      systemClock = clockAt(t_s),
+      eventTimeColumn = "event_time",
       eventTime = t_e
     ).merge(None, curr)
       .as[EmployeeEvent]
-      .orderBy("system_time", "event_time", "id")
+      .orderBy("event_time", "id")
 
     val expected = sc
       .parallelize(
         Seq(
-          EmployeeEvent(t_s, t_e, "I", 1, "Alice", 100),
-          EmployeeEvent(t_s, t_e, "I", 2, "Bob", 80),
-          EmployeeEvent(t_s, t_e, "I", 3, "Charlie", 120)
+          EmployeeEvent(t_e, "I", 1, "Alice", 100),
+          EmployeeEvent(t_e, "I", 2, "Bob", 80),
+          EmployeeEvent(t_e, "I", 3, "Charlie", 120)
         )
       )
       .toDS
-      .orderBy("system_time", "event_time", "id")
+      .orderBy("event_time", "id")
 
     assertDatasetEquals(expected, actual)
   }
@@ -106,20 +96,18 @@ class MergeStrategySnapshotTest extends FunSuite with KamuDataFrameSuite {
       .toDF()
 
     val t_e1 = new Timestamp(0)
-    val t_s1 = new Timestamp(1)
 
     val prev = new SnapshotMergeStrategy(
       primaryKey = Vector("id"),
-      systemClock = clockAt(t_s1),
+      eventTimeColumn = "event_time",
       eventTime = t_e1
     ).merge(None, curr)
 
     val t_e2 = new Timestamp(2)
-    val t_s2 = new Timestamp(3)
 
     val actual = new SnapshotMergeStrategy(
       primaryKey = Vector("id"),
-      systemClock = clockAt(t_s2),
+      eventTimeColumn = "event_time",
       eventTime = t_e2
     ).merge(Some(prev), curr)
       .as[EmployeeEvent]
@@ -149,35 +137,33 @@ class MergeStrategySnapshotTest extends FunSuite with KamuDataFrameSuite {
       .toDF()
 
     val t_e1 = new Timestamp(0)
-    val t_s1 = new Timestamp(1)
 
     val prev = new SnapshotMergeStrategy(
       primaryKey = Vector("id"),
-      systemClock = clockAt(t_s1),
+      eventTimeColumn = "event_time",
       eventTime = t_e1
     ).merge(None, data1)
 
     val t_e2 = new Timestamp(2)
-    val t_s2 = new Timestamp(3)
 
     val actual = new SnapshotMergeStrategy(
       primaryKey = Vector("id"),
-      systemClock = clockAt(t_s2),
+      eventTimeColumn = "event_time",
       eventTime = t_e2
     ).merge(Some(prev), data2)
       .as[EmployeeEvent]
-      .orderBy("system_time", "event_time", "id")
+      .orderBy("event_time", "id")
 
     val expected = sc
       .parallelize(
         Seq(
-          EmployeeEvent(t_s2, t_e2, "D", 1, "Alice", 100),
-          EmployeeEvent(t_s2, t_e2, "U", 3, "Charlie", 130),
-          EmployeeEvent(t_s2, t_e2, "I", 4, "Dan", 100)
+          EmployeeEvent(t_e2, "D", 1, "Alice", 100),
+          EmployeeEvent(t_e2, "U", 3, "Charlie", 130),
+          EmployeeEvent(t_e2, "I", 4, "Dan", 100)
         )
       )
       .toDS()
-      .orderBy("system_time", "event_time", "id")
+      .orderBy("event_time", "id")
 
     assertDatasetEquals(expected, actual)
   }
@@ -206,37 +192,35 @@ class MergeStrategySnapshotTest extends FunSuite with KamuDataFrameSuite {
       .toDF()
 
     val t_e1 = new Timestamp(0)
-    val t_s1 = new Timestamp(1)
 
     val prev = new SnapshotMergeStrategy(
       primaryKey = Vector("id"),
-      systemClock = clockAt(t_s1),
+      eventTimeColumn = "event_time",
       eventTime = t_e1
     ).merge(None, data1)
 
     val t_e2 = new Timestamp(2)
-    val t_s2 = new Timestamp(3)
 
     val actual = new SnapshotMergeStrategy(
       primaryKey = Vector("id"),
-      systemClock = clockAt(t_s2),
+      eventTimeColumn = "event_time",
       eventTime = t_e2
     ).merge(Some(prev), data2)
       .as[EmployeeEvent]
-      .orderBy("system_time", "event_time", "id")
+      .orderBy("event_time", "id")
 
     val expected = sc
       .parallelize(
         Seq(
-          EmployeeEvent(t_s2, t_e2, "D", 1, "Alice", 100),
+          EmployeeEvent(t_e2, "D", 1, "Alice", 100),
           // Complete duplicate removed
-          EmployeeEvent(t_s2, t_e2, "U", 3, "Charlie", 130),
+          EmployeeEvent(t_e2, "U", 3, "Charlie", 130),
           // On PK duplicate currently selects first occurrence (undefined behavior in general)
-          EmployeeEvent(t_s2, t_e2, "I", 4, "Dan", 100)
+          EmployeeEvent(t_e2, "I", 4, "Dan", 100)
         )
       )
       .toDS()
-      .orderBy("system_time", "event_time", "id")
+      .orderBy("event_time", "id")
 
     assertDatasetEquals(expected, actual)
   }
@@ -264,14 +248,14 @@ class MergeStrategySnapshotTest extends FunSuite with KamuDataFrameSuite {
 
     val prev = new SnapshotMergeStrategy(
       primaryKey = Vector("id"),
-      systemClock = clockAt(ts(2)),
+      eventTimeColumn = "event_time",
       eventTime = ts(1)
     ).merge(None, data1)
 
     assertThrows[Exception] {
       new SnapshotMergeStrategy(
         primaryKey = Vector("id"),
-        systemClock = clockAt(ts(3)),
+        eventTimeColumn = "event_time",
         eventTime = ts(0)
       ).merge(
         Some(prev),
@@ -302,36 +286,34 @@ class MergeStrategySnapshotTest extends FunSuite with KamuDataFrameSuite {
       .toDF()
 
     val t_e1 = new Timestamp(0)
-    val t_s1 = new Timestamp(1)
 
     val prev = new SnapshotMergeStrategy(
       primaryKey = Vector("id"),
-      systemClock = clockAt(t_s1),
+      eventTimeColumn = "event_time",
       eventTime = t_e1
     ).merge(None, data1)
 
     val t_e2 = new Timestamp(2)
-    val t_s2 = new Timestamp(3)
 
     val actual = new SnapshotMergeStrategy(
       primaryKey = Vector("id"),
-      systemClock = clockAt(t_s2),
+      eventTimeColumn = "event_time",
       eventTime = t_e2
     ).merge(Some(prev), data2)
       .as[EmployeeEventV2]
-      .orderBy("system_time", "event_time", "id")
+      .orderBy("event_time", "id")
 
     val expected = sc
       .parallelize(
         Seq(
-          EmployeeEventV2(t_s2, t_e2, "D", 1, "Alice", null, 100),
-          EmployeeEventV2(t_s2, t_e2, "U", 2, "Bob", "IT", 80),
-          EmployeeEventV2(t_s2, t_e2, "U", 3, "Charlie", "IT", 130),
-          EmployeeEventV2(t_s2, t_e2, "I", 4, "Dan", "Accounting", 100)
+          EmployeeEventV2(t_e2, "D", 1, "Alice", null, 100),
+          EmployeeEventV2(t_e2, "U", 2, "Bob", "IT", 80),
+          EmployeeEventV2(t_e2, "U", 3, "Charlie", "IT", 130),
+          EmployeeEventV2(t_e2, "I", 4, "Dan", "Accounting", 100)
         )
       )
       .toDS()
-      .orderBy("system_time", "event_time", "id")
+      .orderBy("event_time", "id")
 
     assertDatasetEquals(expected, actual)
   }
@@ -358,36 +340,34 @@ class MergeStrategySnapshotTest extends FunSuite with KamuDataFrameSuite {
       .toDF()
 
     val t_e1 = new Timestamp(0)
-    val t_s1 = new Timestamp(1)
 
     val prev = new SnapshotMergeStrategy(
       primaryKey = Vector("id"),
-      systemClock = clockAt(t_s1),
+      eventTimeColumn = "event_time",
       eventTime = t_e1
     ).merge(None, data1)
 
     val t_e2 = new Timestamp(2)
-    val t_s2 = new Timestamp(3)
 
     val actual = new SnapshotMergeStrategy(
       primaryKey = Vector("id"),
-      systemClock = clockAt(t_s2),
+      eventTimeColumn = "event_time",
       eventTime = t_e2
     ).merge(Some(prev), data2)
       .as[EmployeeEventV2]
-      .orderBy("system_time", "event_time", "id")
+      .orderBy("event_time", "id")
 
     val expected = sc
       .parallelize(
         Seq(
-          EmployeeEventV2(t_s2, t_e2, "D", 1, "Alice", "IT", 100),
-          EmployeeEventV2(t_s2, t_e2, "U", 2, "Bob", null, 80),
-          EmployeeEventV2(t_s2, t_e2, "U", 3, "Charlie", null, 120),
-          EmployeeEventV2(t_s2, t_e2, "I", 4, "Dan", null, 100)
+          EmployeeEventV2(t_e2, "D", 1, "Alice", "IT", 100),
+          EmployeeEventV2(t_e2, "U", 2, "Bob", null, 80),
+          EmployeeEventV2(t_e2, "U", 3, "Charlie", null, 120),
+          EmployeeEventV2(t_e2, "I", 4, "Dan", null, 100)
         )
       )
       .toDS()
-      .orderBy("system_time", "event_time", "id")
+      .orderBy("event_time", "id")
 
     assertDatasetEquals(expected, actual)
   }